initial touchups
